Big epic commit, maybe final
Well not that big and epic, but it consolidates a lot of final polish and tweaking around the reports and their formatting, including some mild overhauls explained under LecturerReport and StudentReport. I may be finished with this now (not that the deadline leaves much choice, but will test it some more). Hope there are no glaring catastrophes hiding under my nose. I left in some output files to try to prove the generation is working on my end. This induced the odd night terror but was more fun than not and very educational. Thanks again Sam for the classes!