The End
The End! Had to modify the Lecturer query again. It outputs properly at time of writing. Who knows whether it will all explode if I tweak anything else. Finished with this now for better or worse. Learned and failed and tried a lot. Thanks for reading this head melted stream of consciousness and look after yourself!